+++ core/controller/shopkeeper.php
@@ -79,9 +79,7 @@ class shopkeeperDocLister extends site_contentDocLister
                         $item = $extUser->setUserData($item); //[+user.id.createdby+], [+user.fullname.publishedby+], [+dl.user.publishedby+]....
                     }
 
-                    if ($extSummary) {
-                        $item['summary'] = $extSummary->init($this, array("content" => $item['content'], "summary" => $this->getCFGDef("summary", "")));
-                    }
+                    $item['summary'] = $extSummary ? $this->getSummary($item, $extSummary) : '';
 
                     $item = array_merge($item, $sysPlh); //inside the chunks available all placeholders set via $modx->toPlaceholders with prefix id, and with prefix sysKey